SDK-1542: Organise check report files
