+++ lib/yoti_sandbox/doc_scan.rb
@@ -4,15 +4,15 @@ require_relative 'doc_scan/client'
 require_relative 'doc_scan/request/task_results'
 require_relative 'doc_scan/request/check_reports'
 require_relative 'doc_scan/request/response_config'
-require_relative 'doc_scan/request/breakdown'
-require_relative 'doc_scan/request/recommendation'
-require_relative 'doc_scan/request/detail'
-require_relative 'doc_scan/request/check'
-require_relative 'doc_scan/request/document_check'
-require_relative 'doc_scan/request/document_authenticity_check'
-require_relative 'doc_scan/request/document_face_match_check'
-require_relative 'doc_scan/request/document_text_data_check'
-require_relative 'doc_scan/request/document_text_data_extraction_task'
-require_relative 'doc_scan/request/liveness_check'
-require_relative 'doc_scan/request/zoom_liveness_check'
 require_relative 'doc_scan/request/document_filter'
+require_relative 'doc_scan/request/check/check'
+require_relative 'doc_scan/request/check/document_check'
+require_relative 'doc_scan/request/check/document_authenticity_check'
+require_relative 'doc_scan/request/check/document_face_match_check'
+require_relative 'doc_scan/request/check/document_text_data_check'
+require_relative 'doc_scan/request/check/liveness_check'
+require_relative 'doc_scan/request/check/zoom_liveness_check'
+require_relative 'doc_scan/request/check/report/breakdown'
+require_relative 'doc_scan/request/check/report/recommendation'
+require_relative 'doc_scan/request/check/report/detail'
+require_relative 'doc_scan/request/task/document_text_data_extraction_task'
