+++ lib/yoti_sandbox/doc_scan/request/check/document_text_data_check.rb
