+++ lib/yoti_sandbox/doc_scan/request/check/liveness_check.rb
