+++ lib/yoti_sandbox/doc_scan/request/check/report/breakdown.rb
