+++ lib/yoti_sandbox/doc_scan/request/task/document_text_data_extraction_task.rb
