+++ spec/yoti_sandbox/doc_scan/request/check/check_spec.rb
