+++ spec/yoti_sandbox/doc_scan/request/check/document_authenticity_check_spec.rb
