+++ spec/yoti_sandbox/doc_scan/request/check/report/breakdown_spec.rb
