+++ spec/yoti_sandbox/doc_scan/request/check/report/detail_spec.rb
