+++ spec/yoti_sandbox/doc_scan/request/check/report/recommendation_spec.rb
